+++ src/main/java/dta/sfmflow/datagen/ModBlockTagProvider.java
@@ -27,15 +27,16 @@ public class ModBlockTagProvider extends BlockTagsProvider {
 				ModBlocks.HARDENED_CABLE_BLOCK.get(), ModBlocks.REDSTONE_EMITTER_BLOCK.get(),
 				ModBlocks.REDSTONE_RECEIVER_BLOCK.get(), ModBlocks.OBSERVER_CABLE_BLOCK.get(),
 				ModBlocks.ITEM_EJECTOR_VALVE_BLOCK.get(), ModBlocks.ITEM_VACUUM_VALVE_BLOCK.get(),
-				ModBlocks.FLUID_HATCH_CABLE_BLOCK.get(), ModBlocks.CABLE_CLUSTER_BLOCK.get(),
-				ModBlocks.ADVANCED_CABLE_CLUSTER_BLOCK.get());
+				ModBlocks.FLUID_EJECTOR_VALVE_BLOCK.get(), ModBlocks.FLUID_VACUUM_VALVE_BLOCK.get(),
+				ModBlocks.CABLE_CLUSTER_BLOCK.get(), ModBlocks.ADVANCED_CABLE_CLUSTER_BLOCK.get());
 
 		tag(ModTags.CABLES).add(ModBlocks.CABLE_BLOCK.get(), ModBlocks.HARDENED_CABLE_BLOCK.get(),
 				ModBlocks.REDSTONE_EMITTER_BLOCK.get(), ModBlocks.REDSTONE_RECEIVER_BLOCK.get(),
 				ModBlocks.OBSERVER_CABLE_BLOCK.get(), ModBlocks.ITEM_EJECTOR_VALVE_BLOCK.get(),
-				ModBlocks.ITEM_VACUUM_VALVE_BLOCK.get(), ModBlocks.FLUID_HATCH_CABLE_BLOCK.get(),
-				ModBlocks.CABLE_CLUSTER_BLOCK.get(), ModBlocks.ADVANCED_CABLE_CLUSTER_BLOCK.get());
-
+				ModBlocks.ITEM_VACUUM_VALVE_BLOCK.get(), ModBlocks.FLUID_EJECTOR_VALVE_BLOCK.get(),
+				ModBlocks.FLUID_VACUUM_VALVE_BLOCK.get(), ModBlocks.CABLE_CLUSTER_BLOCK.get(),
+				ModBlocks.ADVANCED_CABLE_CLUSTER_BLOCK.get());
+		
 		tag(ModTags.REDSTONE_CABLES).add(ModBlocks.REDSTONE_EMITTER_BLOCK.get(),
 				ModBlocks.REDSTONE_RECEIVER_BLOCK.get(), ModBlocks.OBSERVER_CABLE_BLOCK.get());